[修复整数溢出和代码格式优化]: 增强内存分配安全性并调整代码风格
- **修复整数溢出问题**: 在`GifEncoder.cpp`和`gif.h`的多处内存分配中，将`int`类型变量显式转换为`size_t`，避免大尺寸图像处理时可能导致的整数溢出。
- **优化内存分配安全性**: 使用`static_cast<size_t>`处理`width * height`等计算，确保内存分配参数在32位/64位系统下均有效。
- **调整代码格式**: 统一运算符空格风格（如`*`前后增加空格），提升代码可读性。
- **修复潜在错误**: 修正`GifPartition`函数中`pivotValue`的计算表达式格式，消除可能的逻辑歧义。
- **提高兼容性**: 在`GifMakePalette`和`GifBegin`函数中，显式处理图像尺寸计算，避免内存分配不足或越界问题。

+++ src/3rdparty/gif/gif-h/gif.h
@@ -165,9 +165,10 @@ void GifSwapPixels(uint8_t *image, int pixA, int pixB)
 }
 
 // just the partition operation from quicksort
-auto GifPartition(uint8_t *image, const int left, const int right, const int elt, int pivotIndex) -> int
+auto GifPartition(uint8_t *image, const int left, const int right, const int elt, int pivotIndex)
+    -> int
 {
-    const int pivotValue = image[(pivotIndex) *4 + elt];
+    const int pivotValue = image[(pivotIndex) * 4 + elt];
     GifSwapPixels(image, pivotIndex, right - 1);
     int storeIndex = left;
     bool split = 0;
@@ -383,7 +384,7 @@ void GifMakePalette(const uint8_t *lastFrame,
 
     // SplitPalette is destructive (it sorts the pixels by color) so
     // we must create a copy of the image for it to destroy
-    size_t imageSize = (size_t) (width * height * 4 * sizeof(uint8_t));
+    size_t imageSize = static_cast<size_t>(width) * height * 4 * sizeof(uint8_t);
     uint8_t *destroyableImage = (uint8_t *) GIF_TEMP_MALLOC(imageSize);
     memcpy(destroyableImage, nextFrame, imageSize);
 
@@ -431,7 +432,7 @@ void GifDitherImage(const uint8_t *lastFrame,
 
     for (int ii = 0; ii < numPixels * 4; ++ii) {
         uint8_t pix = nextFrame[ii];
-        int32_t pix16 = (int32_t) (pix) *256;
+        int32_t pix16 = (int32_t) (pix) * 256;
         quantPixels[ii] = pix16;
     }
 
@@ -787,7 +788,7 @@ auto GifBegin(GifWriter *writer,
     writer->firstFrame = true;
 
     // allocate
-    writer->oldImage = (uint8_t *) GIF_MALLOC(width * height * 4);
+    writer->oldImage = (uint8_t *) GIF_MALLOC(static_cast<size_t>(width) * height * 4);
 
     fputs("GIF89a", writer->f);
